+++ crates/parol-ls/src/parol_ls_grammar_trait.rs
@@ -250,27 +250,27 @@ pub struct Declaration3 {
 ///
 /// Type derived for production 11
 ///
-/// ScannerDirectives: "%line_comment" : OwnedToken String Comments;
+/// ScannerDirectives: "%line_comment" : OwnedToken TokenLiteral Comments;
 ///
 #[allow(dead_code)]
 #[derive(Builder, Debug, Clone)]
 pub struct ScannerDirectives0 {
     pub percent_line_underscore_comment: crate::parol_ls_grammar::OwnedToken, /* %line_comment */
-    pub string: Box<String>,
+    pub token_literal: Box<TokenLiteral>,
     pub comments: Box<Comments>,
 }
 
 ///
 /// Type derived for production 12
 ///
-/// ScannerDirectives: "%block_comment" : OwnedToken String String Comments;
+/// ScannerDirectives: "%block_comment" : OwnedToken TokenLiteral TokenLiteral Comments;
 ///
 #[allow(dead_code)]
 #[derive(Builder, Debug, Clone)]
 pub struct ScannerDirectives1 {
     pub percent_block_underscore_comment: crate::parol_ls_grammar::OwnedToken, /* %block_comment */
-    pub string: Box<String>,
-    pub string0: Box<String>,
+    pub token_literal: Box<TokenLiteral>,
+    pub token_literal0: Box<TokenLiteral>,
     pub comments: Box<Comments>,
 }
 
@@ -1419,13 +1419,13 @@ impl<'t, 'u> ParolLsGrammarAuto<'t, 'u> {
 
     /// Semantic action for production 11:
     ///
-    /// ScannerDirectives: "%line_comment" : OwnedToken String Comments;
+    /// ScannerDirectives: "%line_comment" : OwnedToken TokenLiteral Comments;
     ///
     #[parol_runtime::function_name::named]
     fn scanner_directives_0(
         &mut self,
         percent_line_underscore_comment: &ParseTreeStackEntry<'t>,
-        _string: &ParseTreeStackEntry<'t>,
+        _token_literal: &ParseTreeStackEntry<'t>,
         _comments: &ParseTreeStackEntry<'t>,
         parse_tree: &Tree<ParseTreeType<'t>>,
     ) -> Result<()> {
@@ -1436,10 +1436,10 @@ impl<'t, 'u> ParolLsGrammarAuto<'t, 'u> {
             .try_into()
             .into_diagnostic()?;
         let comments = pop_item!(self, comments, Comments, context);
-        let string = pop_item!(self, string, String, context);
+        let token_literal = pop_item!(self, token_literal, TokenLiteral, context);
         let scanner_directives_0_built = ScannerDirectives0Builder::default()
             .percent_line_underscore_comment(percent_line_underscore_comment)
-            .string(Box::new(string))
+            .token_literal(Box::new(token_literal))
             .comments(Box::new(comments))
             .build()
             .into_diagnostic()?;
@@ -1457,14 +1457,14 @@ impl<'t, 'u> ParolLsGrammarAuto<'t, 'u> {
 
     /// Semantic action for production 12:
     ///
-    /// ScannerDirectives: "%block_comment" : OwnedToken String String Comments;
+    /// ScannerDirectives: "%block_comment" : OwnedToken TokenLiteral TokenLiteral Comments;
     ///
     #[parol_runtime::function_name::named]
     fn scanner_directives_1(
         &mut self,
         percent_block_underscore_comment: &ParseTreeStackEntry<'t>,
-        _string: &ParseTreeStackEntry<'t>,
-        _string0: &ParseTreeStackEntry<'t>,
+        _token_literal: &ParseTreeStackEntry<'t>,
+        _token_literal0: &ParseTreeStackEntry<'t>,
         _comments: &ParseTreeStackEntry<'t>,
         parse_tree: &Tree<ParseTreeType<'t>>,
     ) -> Result<()> {
@@ -1475,12 +1475,12 @@ impl<'t, 'u> ParolLsGrammarAuto<'t, 'u> {
             .try_into()
             .into_diagnostic()?;
         let comments = pop_item!(self, comments, Comments, context);
-        let string0 = pop_item!(self, string0, String, context);
-        let string = pop_item!(self, string, String, context);
+        let token_literal0 = pop_item!(self, token_literal0, TokenLiteral, context);
+        let token_literal = pop_item!(self, token_literal, TokenLiteral, context);
         let scanner_directives_1_built = ScannerDirectives1Builder::default()
             .percent_block_underscore_comment(percent_block_underscore_comment)
-            .string(Box::new(string))
-            .string0(Box::new(string0))
+            .token_literal(Box::new(token_literal))
+            .token_literal0(Box::new(token_literal0))
             .comments(Box::new(comments))
             .build()
             .into_diagnostic()?;